# Debug mode performance improvements