--- a/Sources/LZ77/HuffmanCoding/LZ77.Distance.swift
+++ b/Sources/LZ77/HuffmanCoding/LZ77.Distance.swift
@@ -21,10 +21,38 @@ extension LZ77.Distance
 {
     var decade:Int
     {
-        .init(self.storage & 0x00ff)
+        let value = self.storage & 0x00ff
+        #if DEBUG
+            // this hacky integer conversion makes this function about 17x faster in
+            // debug mode. cause of Int, we have to handle both 32-bit and 64-bit
+            // systems separately.
+            if MemoryLayout<Int>.stride == 8 {
+                let tuple:(UInt16, UInt16, UInt16, UInt16) = (value, 0, 0, 0)
+                return unsafeBitCast(tuple, to: Int.self)
+            } else {
+                let tuple:(UInt16, UInt16) = (value, 0)
+                return unsafeBitCast(tuple, to: Int.self)
+            }
+        #else
+            return .init(value)
+        #endif
     }
     var length:Int
     {
-        .init(self.storage >> 8)
+        let value = self.storage >> 8
+        #if DEBUG
+            // this hacky integer conversion makes this function about 24x faster in
+            // debug mode. cause of Int, we have to handle both 32-bit and 64-bit
+            // systems separately.
+            if MemoryLayout<Int>.stride == 8 {
+                let tuple:(UInt16, UInt16, UInt16, UInt16) = (value, 0, 0, 0)
+                return unsafeBitCast(tuple, to: Int.self)
+            } else {
+                let tuple:(UInt16, UInt16) = (value, 0)
+                return unsafeBitCast(tuple, to: Int.self)
+            }
+        #else
+            return .init(value)
+        #endif
     }
 }