--- a/Sources/LZ77/Inflator/LZ77.InflatorTables.swift
+++ b/Sources/LZ77/Inflator/LZ77.InflatorTables.swift
@@ -106,7 +106,26 @@ extension LZ77.InflatorTables
         // now i know why everyone at apple asks this same coding interview question
         self.storage.withUnsafeMutablePointerToElements
         {
-            .init($0[.init(byte)])
+            #if DEBUG
+                // these hacky integer conversions make this function about 7x faster in
+                // debug mode. cause of Int, we have to handle both 32-bit and 64-bit
+                // systems separately.
+                if MemoryLayout<Int>.stride == 8 {
+                    let indexTuple:(UInt16, UInt16, UInt16, UInt16) = (byte, 0, 0, 0)
+                    let index = unsafeBitCast(indexTuple, to: Int.self)
+                    let value = $0[index]
+                    let valueTuple:(UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8) = (value, 0, 0, 0, 0, 0, 0, 0)
+                    return unsafeBitCast(valueTuple, to: Int.self)
+                } else {
+                    let indexTuple:(UInt16, UInt16) = (byte, 0)
+                    let index = unsafeBitCast(indexTuple, to: Int.self)
+                    let value = $0[index]
+                    let valueTuple:(UInt8, UInt8, UInt8, UInt8) = (value, 0, 0, 0)
+                    return unsafeBitCast(valueTuple, to: Int.self)
+                }
+            #else
+                .init($0[.init(byte)])
+            #endif
         }
     }
     private