@stackotter 2024-09-25
As some additional context. I've got a test app written with SwiftCrossUI that has 12 500x500 images (among other things). Originally they were JPGs, but the swift `jpeg` library takes over 20 seconds to load all of those images in debug mode (and still about 3 seconds in release mode). I converted them to PNGs, and `swift-png` took around 4 seconds to load them all in debug mode. And then I converted them to WEBPs, and they loaded basically instantly in both debug and release mode (I'm using `libwebp` written in C).

I'm thinking that I might update my `swift-image-formats` library to use C libraries for jpg and png by default with an option to use `swift-png` and `jpeg` for ultimate portability by setting an environment variable (or a feature flag if/when SwiftPM gets those). And then there won't be as much pressure on improving the debug mode performance of `swift-png` and `jpeg` in the near future.