@stackotter 2025-01-15
I've switched `swift-image-formats` over to `libpng` and I feel like this PR will take a lot of effort to fragment and integrate incrementally, so I'm closing this for now. But if anyone is attempting to improve `swift-png`'s debug mode performance in the future, this should at least provide a good starting point!