@tayloraswift 2025-01-15
i’m very sad to hear that, but it’s definitely understandable. it’s quite ridiculous that in 2025, SwiftPM still cannot compile dependencies in release mode separately from the app being developed.

like i said before, manually replicating compiler optimizations in library source code to work around a limitation of one particular build system is not a sustainable approach.
i do hope though when SwiftPM catches up to industry standard build systems, you’ll give swift-png another shot.