@stackotter 2025-01-15
> it’s quite ridiculous that in 2025, SwiftPM still cannot compile dependencies in release mode separately from the app being developed.

Yeah I agree, it's quite ridiculous. It gives closed source SDKs quite a few advantages over open source packages, which is not how it should be!

Replacing Apple's system libraries with cross-platform pure Swift libraries shouldn't require giving up debug mode 😢

Perhaps I should pitch it at some point in the near future.