@S1D1T1 2025-03-14
Thank you for this thread which solves my mystery. Turns out I've been driving around the last couple weeks with the parking brake stuck on. Recompiled for Release and vroom.